+++ src/executeCommands.ts
@@ -1,6 +1,6 @@
 import { exec as execute } from "https://deno.land/x/execute@v1.1.0/mod.ts"
-import { exec } from "https://deno.land/x/exec/mod.ts"
-import { promptString } from "./utils.ts"
+import { exec } from "https://deno.land/x/exec@0.0.5/mod.ts"
+import { switchBackToMainBranch } from "./utils.ts"
 
 export const executeCommands = async (prTitle: string, branchName: string, prBody: string) => {
   await exec(`bit checkout -b ${branchName}`)
@@ -19,7 +19,7 @@ export const executeCommands = async (prTitle: string, branchName: string, prBod
   console.log(`Pushing ${branchName}...`)
   const {
     status: { success: pushSuccess },
-  } = await exec(`git push --set-upstream origin ${branchName} --no-verify`)
+  } = await exec(`git push --set-upstream origin ${branchName}`)
 
   if (!pushSuccess) {
     console.error("\x1b[31m%s\x1b[0m", "Couldn't push changes.")
@@ -34,13 +34,5 @@ export const executeCommands = async (prTitle: string, branchName: string, prBod
 
   await exec("gh pr view -w")
 
-  const allBranches = await execute("git branch -a")
-  const mainBranch = allBranches.includes("master") ? "master" : "main"
-
-  const switchBackToMainBranch = await promptString(`Switch back to ${mainBranch}? (y,N) `)
-
-  if (switchBackToMainBranch === "y") {
-    console.log(`Switching to ${mainBranch}`)
-    await exec(`git checkout ${mainBranch}`)
-  }
+  await switchBackToMainBranch()
 }
